implement  discovery for ESP32 USB Serial JTAG on Linux and Windows
