+++ go/internal/shared/discovery/serial_common.go
@@ -1,0 +1,30 @@
+package discovery
+
+import (
+	"fmt"
+	"time"
+)
+
+// SerialDevice holds a serial port path and its USB connection time.
+type SerialDevice struct {
+	Port           string
+	ConnectionTime time.Time
+}
+
+// ResolveESP32SerialPort returns the most recently connected ESP32 serial port.
+func ResolveESP32SerialPort() (string, error) {
+	devices, err := ResolveESP32SerialPorts()
+	if err != nil {
+		return "", err
+	}
+	if len(devices) == 0 {
+		return "", fmt.Errorf("no ESP32 serial port found")
+	}
+	best := devices[0]
+	for _, d := range devices[1:] {
+		if d.ConnectionTime.After(best.ConnectionTime) {
+			best = d
+		}
+	}
+	return best.Port, nil
+}
