+++ go/internal/shared/discovery/serial_linux.go
@@ -11,17 +11,19 @@ import (
 	"github.com/wendylabsinc/wendy/go/internal/shared/models"
 )
 
-// ResolveESP32SerialPort finds the serial port for an ESP32-C6 device on Linux.
-// It walks /sys/class/tty/ttyACM* and matches the Espressif VID/PID.
-func ResolveESP32SerialPort() (string, error) {
+// ResolveESP32SerialPorts returns all connected serial ports whose USB VID/PID
+// match the ESP32 constants, along with each device node's modification time as
+// a proxy for when the device was plugged in.
+func ResolveESP32SerialPorts() ([]SerialDevice, error) {
 	entries, err := filepath.Glob("/sys/class/tty/ttyACM*")
 	if err != nil {
-		return "", fmt.Errorf("globbing tty entries: %w", err)
+		return nil, fmt.Errorf("globbing tty entries: %w", err)
 	}
 
 	wantVID := strings.TrimPrefix(models.ESP32VendorID, "0x")
 	wantPID := strings.TrimPrefix(models.ESP32ProductID, "0x")
 
+	var result []SerialDevice
 	for _, entry := range entries {
 		deviceSymlink := filepath.Join(entry, "device")
 		resolvedIface, err := filepath.EvalSymlinks(deviceSymlink)
@@ -33,24 +35,26 @@ func ResolveESP32SerialPort() (string, error) {
 			continue
 		}
 		usbDevPath := filepath.Dir(resolvedIface)
-		vidPath := filepath.Join(usbDevPath, "idVendor")
-		pidPath := filepath.Join(usbDevPath, "idProduct")
 
-		vid, err := os.ReadFile(vidPath)
+		vid, err := os.ReadFile(filepath.Join(usbDevPath, "idVendor"))
 		if err != nil {
 			continue
 		}
-		pid, err := os.ReadFile(pidPath)
+		pid, err := os.ReadFile(filepath.Join(usbDevPath, "idProduct"))
 		if err != nil {
 			continue
 		}
 
-		if strings.TrimSpace(string(vid)) == wantVID &&
-			strings.TrimSpace(string(pid)) == wantPID {
-			devName := filepath.Base(entry)
-			return "/dev/" + devName, nil
+		if strings.TrimSpace(string(vid)) != wantVID || strings.TrimSpace(string(pid)) != wantPID {
+			continue
 		}
+
+		devPath := "/dev/" + filepath.Base(entry)
+		dev := SerialDevice{Port: devPath}
+		if info, statErr := os.Stat(devPath); statErr == nil {
+			dev.ConnectionTime = info.ModTime()
+		}
+		result = append(result, dev)
 	}
-
-	return "", fmt.Errorf("no ESP32 serial port found (expected /dev/ttyACM* with VID %s)", models.ESP32VendorID)
+	return result, nil
 }
